+++ mythtv/libs/libmythtv/videoout_omx.cpp
@@ -56,7 +56,7 @@ using namespace omxcontext;
 #define LOC QString("VideoOutputOMX: ")
 
 // Roundup a value: y = ROUNDUP(x,4)
-#define ROUNDUP( _x,_z) ((_x) + ((-(int)(_x)) & ((_z) -1)) )
+#define ROUNDUP( _x,_z) ((_x) + ((-(_x)) & ((_z) -1)) )
 
 // VideoFrame <> OMX_BUFFERHEADERTYPE
 #define FRAMESETHDR(f,h) ((f)->priv[3] = reinterpret_cast<unsigned char* >(h))
